>>> Freontend: When creating dates, allow yearly dates to be used where a numeric year input is expected

+++ dates/convert.m
@@ -60,7 +60,7 @@ else
 end
 
 [flag, toFreq] = Frequency.validateFrequency(toFreq);
-outputDate = numeric.convert(inputDate, double(toFreq), varargin{:});
+outputDate = numeric.convert(double(inputDate), double(toFreq), varargin{:});
 
 if strcmpi(outputClass, 'DateWrapper')
     outputDate = DateWrapper(outputDate);
